pyproject.toml
[tool.poetry]
name = "portifolio"
version = "0.1.0"
description = ""
authors = ["Jose Felipe <felipe.brx.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
django-decouple = "^2.1"
dj-database-url = "^2.1.0"
psycopg2 = "^2.9.9"
django-model-utils = "^4.3.1"
pillow = "^10.0.1"
django-jazzmin = "^2.6.0"
django-ipware = "^5.0.1"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
